refactor: migrate `MessageCreate` event to it's own
Add structure for adding all events to their files
This change is necessary due to the client `intents`.

As each event may or may not be emitted due to missing intents, a better approach is to put each event in a single file to reduce the coupling of all events and to allow unique validation for if the client has permission to handle a required event.